% Generated by roxygen2 (4.0.0): do not edit by hand
\name{conform}
\alias{conform}
\title{Make matrices or arrays conform with each other by adding or removing dimension elements.}
\usage{
conform(x, ..., fill = NA, do.all = c("no", "union", "intersect"),
  along = NULL, excess.ok = TRUE, missing.ok = TRUE)
}
\arguments{
\item{x}{An object or list of objects to conform}

\item{\dots}{The object to conform to, or other objects to conform}

\item{fill}{The fill value (default NA)}

\item{do.all}{If do.all=="union" or "intersect", all arguments are conformed to the
union (or intersection) of all dimnames, and are returned in a list.
With "union", this is the same behavior as when a list is supplied as the
only argument.}

\item{along}{Integer specification of dimensions on which to conform}

\item{excess.ok}{If \code{excess.ok==TRUE}, dimension elements of x not in y are silently deleted.
If \code{excess.ok==FALSE}, an error is generated if x has dimension elements not in y}

\item{missing.ok}{If \code{missing.ok==TRUE}, dimension elements of y not in x are filled with the fill value
If \code{missing.ok==FALSE}, an error is generated if x is missing some dimension elements in y}
}
\description{
Make matrices or arrays conform with each other by adding or removing dimension elements.
}

