% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/is.virtual.array.R
\name{is.virtual.array}
\alias{is.virtual.array}
\alias{is.virtual.array.default}
\alias{is.virtual.array.varray}
\title{Indicate whether an array-like object is native or virtual}
\usage{
is.virtual.array(x)

\method{is.virtual.array}{varray}(x)

\method{is.virtual.array}{default}(x)
}
\arguments{
\item{x}{An array-like object}
}
\value{
A logical value, \code{TRUE} or \code{FALSE}
}
\description{
Indicate whether an array-like object is native or
virtual.  A native array object is a primitive vector
object with a \code{dim} attribute.  A virtual array
object is some object that behaves like an array in that
it has a dim() and subset extraction methods \code{'['}.
This is a generic function with methods defined for class
\code{varray}.
}

