% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Extract.R
\name{lookup.arr}
\alias{lookup.arr}
\title{Lookup values in an array}
\usage{
lookup.arr(x, i, ...)
}
\arguments{
\item{x}{a matrix-like object}

\item{i}{a matrix index}

\item{...}{additional arguments to pass to methods}
}
\description{
Lookup values in an array
}

