% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Extract.R
\name{lookup.arr.sparsetsmat}
\alias{lookup.arr.sparsetsmat}
\title{sparsetsmat method for generic lookup.arr}
\usage{
\method{lookup.arr}{sparsetsmat}(x, i, ..., drop = TRUE)
}
\arguments{
\item{x}{a sparsetsmat object}

\item{i}{a matrix index}

\item{...}{unused}

\item{drop}{should dimensions be dropped from a vector result}
}
\description{
sparsetsmat method for generic lookup.arr
}
\note{
The lookup.arr.sparsetsmat method is an alias for the standard extract method
}

