% Generated by roxygen2 (4.0.0): do not edit by hand
\name{M}
\alias{M}
\title{Convert to ordinary matrix or array}
\usage{
M(x)
}
\arguments{
\item{x}{the object to convert}
}
\value{
An atomic array version of \code{x}
}
\description{
Shorthand for as.matrix or as.array
}
\details{
If \code{x} is an atomic object and is not a virtual array, returns
 \code{x}, otherwise returns \code{as.matrix(x)} if \code{x} has two
 dimensions or \code{as.array(x)} otherwise.
}
\examples{
x <- varray(cbind(A=c(a=1)), cbind(B=c(b=2)))
x
M(x)
as.array(x, fill=0)
as.matrix(x)
M(x) + 10
}
\seealso{
\code{link{as.array}} \code{link{as.matrix}}
}

