% Generated by roxygen2 (4.0.0): do not edit by hand
\name{pdfmark}
\alias{pdfmark}
\title{Generate a single PDF bookmark, optionally writing a page number on the current device}
\usage{
pdfmark(title, marks = list(), page = max(marks$page, 0) + inc, inc = 1,
  level = 1, open = level <= 2, plot.it = FALSE, pos = c("bottom",
  "bottomleft", "bottomright", "top", "topleft", "topright"),
  line = attr(marks, "plot.pageno")$line, adj = attr(marks,
  "plot.pageno")$adj, outer = non.null(attr(marks, "plot.pageno")$outer,
  TRUE), format = non.null(attr(marks, "plot.pageno")$format, "Page \%s"),
  mtext.args = non.null(attr(marks, "plot.pageno")$args, list()))
}
\arguments{
\item{title}{text of the bookmark}

\item{marks}{a dataframe of pdf bookmarks (a return value from this function)}

\item{page}{the page number to use, both in the bookmark and to write on the device.  Defaults to incrementing the page number associated with the \code{marks} object.}

\item{inc}{the automatic increment to the page number.  Supply as zero when adding another bookmark for the same page.}

\item{level}{nesting level for the bookmark; top level is 1, bottom level is 4}

\item{open}{when the PDF file is viewed, are the sub-levels initially shown open or closed?}

\item{plot.it}{should the page number be written on the plotting device?}

\item{pos}{controls writing the page number to the plotting device, one of 'bottom', 'bottomleft', 'bottomright', 'top', 'topleft', 'topright'}

\item{line}{position of the page number on the plotting device}

\item{adj}{adjustment for the position of the page number to the plotting device.  A number between 0 (left adjusted) and 1 (right adjusted)}

\item{outer}{(logical) should the page number be written on the plotting device in the outer margins? Default is TRUE.}

\item{format}{controls writing the page number to the plotting device; formatted using \code{sprintf()}. Default is \code{'Page \%s'}.}

\item{mtext.args}{other arguments to \code{mtext()} to use when writing the page number to the plotting device}
}
\value{
A dataframe of pdf bookmarks, suitable for passing to \code{\link{pdf.pdfmark}}.  This function can also have the side effect of writing a page number on the plotting device.
}
\description{
This function records data for one PDF bookbark.  It can
also keep track of the current page number, and can
optionally write a page number on the current plotting
device.  The function \code{\link{pdf.pdfmark}} adds bookmarks to an existing PDF file.
}
\details{
The controls for writing the page number to the plotting device are stored in the return value and reused for the next call to pdfmark()
}
\seealso{
\code{\link{pdf.pdfmark}} for adding bookmarks to a PDF file
}

