% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Replace.R
\name{[<-.sparsetsmat}
\alias{[<-.sparsetsmat}
\title{Replace.sparsetsmat}
\usage{
\method{[}{sparsetsmat}(x, i, j, ...,
  verbose = getOption("sparsetsmat.verbose", FALSE)) <- value
}
\arguments{
\item{x}{a sparsetsmat object}

\item{i}{the row indices, can be numeric, character, Date
or POSIXct, or a 2-column matrix or dataframe (or list)
for matrix indexing.}

\item{j}{the j indices, can be numeric, character or NULL
in the case of matrix indexing}

\item{...}{not used}

\item{verbose}{if TRUE, print informative messages}

\item{value}{the replacement values}
}
\description{
Replace an ordinary sub-matrix or values in a sparsetsmat object.
}

