\name{Rigroup}
\alias{Rigroup}
\concept{Rigroup package}
\concept{grouping}
\concept{subsets}
\concept{by}
\title{package to calculate selected basic statistics for each group }
\description{
  This package provides a number of optimized functions to quickly
  calculate a selected set of basic functions for each group in a
  data vector x where a small integer group vector i indicates
  group membership.  Functions include Alls, Anys, Counts, Maxs,
  Means, Mins, Prods, Ranges and Sums.
}
\seealso{
  \code{\link{igroupAlls}}
  \code{\link{igroupAnys}}
  \code{\link{igroupCounts}}
  \code{\link{igroupMaxs}}
  \code{\link{igroupMeans}}
  \code{\link{igroupMins}}
  \code{\link{igroupProds}}
  \code{\link{igroupRanges}}
  \code{\link{igroupSums}}
}
\references{ None }
\author{K. Hendricks with lots of help from B. Dunlap}
\keyword{package}
