% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{package}
\name{sparsetsmat-package}
\alias{sparsetsmat-package}
\title{Sparse storage for persistent time-series data}
\description{
The \code{sparsetsmat} class provides an object to store
slow changing time-series data in a compact way.
}
\details{
The data is stored compactly by assuming that if no data is
stored for a particular id on a particular date, the most
recent data available for that id should be used.  The
format is designed for fast extraction, but not for fast
modification or addition of new data.

A \code{sparsetsmat} object behaves like an ordinary
matrix with respect to many methods, including extraction
(subsetting using the \code{x[i,j]} syntax), \code{dim},
\code{dimnames}, etc.  Rownames are dates (stored as
either \code{Date} or \code{POSIXct} objects) and
colnames are arbitrary character labels.  The data type
stored can be any atomic R object (e.g., double, integer,
character, etc.)

For example, a dense matrix with the following structure
\tabular{rrrr}{
             \tab X \tab Y \tab Z \cr
'2001-01-01' \tab 'a' \tab NA \tab NA \cr
'2001-01-03' \tab 'a' \tab 'b' \tab NA \cr
'2001-01-05' \tab 'a' \tab 'B' \tab 'C'
}
is stored with effectively the following data:
\tabular{rrr}{
date \tab id \tab value \cr
'2001-01-01' \tab 'X' \tab 'a' \cr
'2001-01-03' \tab 'Y' \tab 'b' \cr
'2001-01-05' \tab 'Y' \tab 'B' \cr
'2001-01-05' \tab 'Z' \tab 'C'
}

The result of subsetting a \code{sparsetsmat} object is
always an ordnary dense matrix.

The indices supplied to the subset method can be
character, \code{Date}, or \code{POSIXct}.  Unlike with
ordinary matrices, the subsetting method for
\code{sparsetsmat} will automatically interpolate data
for dates in the date range of the object.

Subsetting is intended to be as efficient as possible and
is coded in C++ using Rcpp.  Relative to subsetting a
large ordinary matrix, subsetting a \code{sparsetsmat}
object takes between 1 and 6 times as long.  See the
source package file \code{tests/zhuge.Rt} for timing examples.

\code{sparsetsmat} objects differ from the sparse matrix
objects of the \code{Matrix} package in that the uncoded
value for \code{sparsetsmat} objects is the most recent
value in the column, or \code{NA} if there is no prior
value, whereas in the \code{Matrix} package, the uncoded
value is zero.  Furthermore, the \code{sparsetsmat}
package does not provide any matrix-algebra routines that
work directly on the compactly stored data.  The
\code{sparsetsmat} package does only one thing: store
slow-changing persistent time series data in a compact
manner, and convert that data to dense matrices in a
reasonably efficient manner.
}
\seealso{
\code{\link{sparsetsmat}}
}

