% Generated by roxygen2 (4.0.0): do not edit by hand
\name{sparsetsmat}
\alias{sparsetsmat}
\title{sparsetsmat}
\usage{
sparsetsmat(x, ...)
}
\arguments{
\item{drop.unneeded.dates}{Logical.  If TRUE, dates where
no value changes are completely forgotten.  If FALSE,
dates where no values change are still recorded in the
\code{all.dates} component of the object.  In both cases
dates where no value changes do not appear in the
\code{date} component of the returned object.  In both
cases, \code{all.dates} is a sorted version of the dates.}
}
\value{
A \code{sparsetsmat} (S3) object with the following components:
\itemize{

  \item dates The dates of the underlying data.  Can be
\code{Date}, \code{POSIXct}, \code{double} or
\code{integer}

  \item values The values of the underlying data.  Can be
any atomic data type.  It is necessary to be able to do
operations like \code{matrix(x$values[1:6], ncol=2}) on
the values.  This vector must be the same length as
\code{date}.

  \item ids The id values.  May be sorted.

  \item id.idx The start value index into \code{date} and
\code{values} for each id.  Can contain NA values where an
id does not occur in the data.

  \item id.noc The number of occurences of each id in
\code{dates} and \code{values}.  Can contain zeros.

  \item all.dates class numeric, Date or POSIXct.  Will
  be sorted unless drop.unneeded.dates was FALSE and x was
  a matrix with out-of-order dates as rownames.

   \item df.colnames colnames of the data.frame version

  \item backfill logical
}

The data are grouped by \code{ids} (possibly sorted) first
and sorted by \code{dates} second.  If there are no values
for a particular id the value in \code{id.idx} is
ignored.
}
\description{
Sparse storage of persistent time-series data
}
\details{
A sparsetsmat object is stored as a list of four
vectors: the date index, the identifier values and
indices, and the data values.  The identifiers correspond
to the columns in matrix data.  The date index and the
values are vectors of the same length.  The identifier
indices are the starting indices for each identifier in \code{id.idx}.
}

