% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/as.matrix.sparsetsmat.R, R/standard.methods.R, R/summary.sparsetsmat.R
\name{as.matrix.sparsetsmat}
\alias{as.array.sparsetsmat}
\alias{as.matrix.sparsetsmat}
\alias{colnames.sparsetsmat}
\alias{dim.sparsetsmat}
\alias{dimnames.sparsetsmat}
\alias{head.sparsetsmat}
\alias{ncol.sparsetsmat}
\alias{nrow.sparsetsmat}
\alias{print.sparsetsmat}
\alias{rownames.sparsetsmat}
\alias{tail.sparsetsmat}
\title{Standard matrix/array methods for sparsetsmat objects}
\usage{
\method{as.matrix}{sparsetsmat}(x, ...)

\method{as.array}{sparsetsmat}(x, ...)

\method{dim}{sparsetsmat}(x)

\method{dimnames}{sparsetsmat}(x)

\method{nrow}{sparsetsmat}(x)

\method{ncol}{sparsetsmat}(x)

\method{rownames}{sparsetsmat}(x)

\method{colnames}{sparsetsmat}(x)

\method{head}{sparsetsmat}(x, n = 6L, ...)

\method{tail}{sparsetsmat}(x, n = 6L, addrownums = TRUE, ...)

\method{print}{sparsetsmat}(x, ...)
}
\arguments{
\item{x}{a sparsetsmat object}

\item{...}{additional arguments}

\item{n}{the number of rows to return}

\item{addrownums}{should row numbers be added if there are none?}
}
\description{
as.matrix: Convert a sparsetsmat object to an ordinary (non-sparse) matrix

as.array: Convert a sparsetsmat object to an ordinary (non-sparse) matrix

dim: return \code{dim} of a sparsetsmat

dimnames: return \code{dimnames} of a sparsetsmat

nrow: return number of rows (virtual) of a sparsetsmat

ncol: return number of cols of a sparsetsmat

rownames: return rownames of a sparsetsmat

colnames: return colnames of a sparsetsmat

head: return the first part of a sparse matrix.

tail: return the last part of a sparse matrix.
}

