% Generated by roxygen2 (4.0.0): do not edit by hand
\name{[.sparsetsmat}
\alias{[.sparsetsmat}
\title{Extract.sparsetsmat}
\usage{
\method{[}{sparsetsmat}(x, i, j, ..., drop = TRUE, vidx = FALSE,
  details = FALSE, backfill = x$backfill, Cpp = TRUE, naked = FALSE)
}
\arguments{
\item{x}{a sparsetsmat object}

\item{i}{the row indices, can be numeric, character, Date
or POSIXct, or a 2-column matrix or dataframe for matrix
indexing.}

\item{j}{the j indices, can be numeric, character or NULL
in the case of matrix indexing}

\item{drop}{logical, default TRUE}

\item{vidx}{logical If TRUE, return integer index of the
unique values, rather than the values themselves.  The
returned integers index into \code{values(x)} (the sorted
unique values in \code{x}).}

\item{details}{logical If TRUE, return a list with the
normalized indices and values.}

\item{backfill}{logical If TRUE, use the oldest data for
earlier dates.}

\item{Cpp}{logical If TRUE, use fast C++ code}

\item{naked}{logical If TRUE do not return dimnames on
the result.  This can result is a very large speeed up
when there are many rownames compared to returned
elements, because constructing rownames involves format
date objects which is slow.}

\item{...}{required by the generic, but extra arguments
not described here are not alowed.}
}
\description{
Extract an ordinary sub-matrix or value from a sparsetsmat object.
}
\details{
Rules for interpreting i: \itemize{

   \item If the time index is numeric (not a Date or
POSIXct), then i is always interpreted as a value, i.e.,
i cannot be interpreted as positional in this case.
E.g., with a numeric time index, an i value of 17 does
not refer to the 17th row, rather it refers to the value
17.  Similarly, with a numeric time index, an i value of
-3 does not result in dropping the 3rd row of the matrix,
rather it refers to a time value of -3 (and negative
numeric time indices can be stored).

   \item If the time index is a date, then if i is
numeric, it is an index (positional), else if i is
character or date, it is a value.  When the time index is
a date, a negative i value has the conventional R
interpretation as dropping that row from the result.

   \item The first column in a matrix index (can be a
dataframe) is interpreted in the same way.

}
}
\note{
Matrix-indexing for \code{sparsetsmat} objects is
slightly more verstatile than for standard matrix objects
in R in that \code{sparsetsmat} objects can take a
dataframe for the matrix index, allowing the use of
different datatypes for specifying the indices on the
rows and columns.
}

