% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/update.varray.ts.R
\name{update.varray.ts}
\alias{update.varray.ts}
\title{Add time series data to a varray object (obselete version of add.tsdata.varray)

Add time series data to a varray object (obselete version
of add.tsdata.varray).  This is not a method for a
generic, though it is documented as one to stop R CMD
check whinging about improper documentation.}
\usage{
\method{update}{varray.ts}(object, data, comp.name = va$comp.name,
  dateblock = "\%Y", format = va$format, along = va$along,
  dimorder = va$dimorder, env.name = va$env.name, envir = NULL,
  naidxok = va$naidxok, change.mode.ok = FALSE,
  keep.ordered = va$keep.ordered, umode = NULL, store.env.name = FALSE,
  fill = NA, ...)
}
\arguments{
\item{object}{The character name of an object to add time-series data to.
The object is changed in place.}

\item{data}{New data to incorporate in the object.
Should have the same number of dimensions as the object
being updated.  Must have dimnames.}

\item{comp.name}{Base of name to use for the component objects in the varray.
Optional, default value is \code{paste('.', objectName, dateblock)}.}

\item{dateblock}{How to translate dates into blocks.  Ignored if argument
\code{comp.name} is supplied. Default value is \code{"\%Y"}.}

\item{format}{Format for parsing dates (which are supplied as dimension names on the
\code{along} dimension).}

\item{along}{Default is 1.}

\item{dimorder}{Default is standard dim order.}

\item{env.name}{Intended to be the name of the environment where the components exist,
not yet fully tested.}

\item{envir}{Intended to be the environment where the components exist,
not yet fully tested.}

\item{naidxok}{Set this attribute on the varray.  Specifies whether the component objects can handle \code{NA} indices.}

\item{change.mode.ok}{Logical.  If TRUE, the mode of new data does not need to match existing data (default FALSE).}

\item{keep.ordered}{Logical.  Specifies which dimensions should be kept ordered.  Can be a
single element or a vector with length equal to the number of
dimensions of \code{object}.}

\item{umode}{Not used (may be used in the future)}

\item{store.env.name}{Not used (may be used in the future)}

\item{fill}{The value to fill missing data with in subset operations}

\item{\dots}{Not used, but needed because \code{update()} is a generic.}
}
\value{
The altered object \code{x}.
}
\description{
Add time series data to a varray object (obselete version of add.tsdata.varray)

Add time series data to a varray object (obselete version
of add.tsdata.varray).  This is not a method for a
generic, though it is documented as one to stop R CMD
check whinging about improper documentation.
}
\details{
Contents of \code{data} are inserted into \code{object}, expanding the
dimensions of \code{object} and creating new component objects if necessary.
\code{update.varray.ts} is an obselete name for \code{add.tsdata.varray}.
}
\seealso{
\code{\link{add.tsdata}}
}

