% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/varray.R, R/vstorage.mode.R
\name{varray.methods}
\alias{[.noretry.varray}
\alias{[.varray}
\alias{[<-.varray}
\alias{as.array.noretry.varray}
\alias{as.array.varray}
\alias{as.matrix.varray}
\alias{dim.varray}
\alias{dim<-.varray}
\alias{dimnames.varray}
\alias{dimnames<-.varray}
\alias{head.varray}
\alias{length<-.varray}
\alias{mode.varray}
\alias{mode<-.varray}
\alias{print.varray}
\alias{storage.mode.varray}
\alias{tail.varray}
\alias{varray.methods}
\alias{vstorage.mode}
\alias{vstorage.mode.varray}
\alias{vstorage.mode<-}
\alias{vstorage.mode<-.varray}
\title{Common matrix and array methods for varray objects}
\usage{
varray.methods()

\method{as.array}{noretry.varray}(x, ..., fill = x$fill)

\method{as.array}{varray}(x, ..., fill = x$fill, maxTries = 30,
  maxWaitSecs = 60, waitSecs = 1, succeedOnTry = NA)

\method{as.matrix}{varray}(x, ..., fill = x$fill)

\method{print}{varray}(x, ...)

\method{dim}{varray}(x)

\method{dimnames}{varray}(x)

\method{mode}{varray}(x)

\method{storage.mode}{varray}(x)

\method{head}{varray}(x, n = 6L, ...)

\method{tail}{varray}(x, n = 6L, addrownums = TRUE, ...)

\method{[}{varray}(x, ..., drop = TRUE, maxTries = 30, maxWaitSecs = 60,
  waitSecs = 1, succeedOnTry = NA, compCheck = TRUE)

\method{[}{noretry.varray}(x, ..., drop = TRUE, compCheck = TRUE)

\method{dimnames}{varray}(x) <- value

\method{dim}{varray}(x) <- value

\method{length}{varray}(x) <- value

\method{mode}{varray}(x) <- value

\method{[}{varray}(x, i, j, ...) <- value

vstorage.mode(x, deep = FALSE)

\method{vstorage.mode}{varray}(x, deep = FALSE)

vstorage.mode(x) <- value

\method{vstorage.mode}{varray}(x) <- value
}
\arguments{
\item{x}{a varray object}

\item{fill}{fill value to use in subsetting}

\item{maxTries}{maximum number of retries waiting for consistency between container and content objects}

\item{maxWaitSecs}{maximum wait before next retry after a failed consistency check (wait times double each iteration until hitting the max)}

\item{waitSecs}{initial wait before next retry after a failed consistency check}

\item{succeedOnTry}{for debugging only: simulate success on this iteration of consistency checking}

\item{n}{the number of rows to return}

\item{addrownums}{should row numbers be added if there are none?}

\item{drop}{should dimensions with extent 1 in the result be dropped?}

\item{compCheck}{should the consistency check for components be done?}

\item{value}{new values to use}

\item{i,j,\dots}{indexing arguments, treated as for ordinary array indexing}

\item{deep}{should individual components be inspected to determine the result?}
}
\description{
Common matrix and array methods for varray objects

Like as.array, but refetch the underlying objects when there are inconsistencies.
Can happen if new versions of the sub-arrays have been resaved before the container object itself is saved
}
\section{Methods (by generic)}{
\itemize{
\item \code{as.array}: Convert a varray to an ordinary array

\item \code{as.array}: Convert a varray to an ordinary array

\item \code{as.matrix}: Convert a varray to an ordinary matrix

\item \code{print}: Print info about a varray object (but not the whole object)

\item \code{dim}: Return dimension size of a varray object

\item \code{dimnames}: Return dimension names of a varray object

\item \code{mode}: Return 'mode' of data in a varray object

\item \code{storage.mode}: Return 'storage.mode' of data in a varray object

\item \code{head}: Return the first part of a varray.

\item \code{tail}: Return the last part of a varray.

\item \code{[}: Returns a subset of a varray objects

\item \code{[}: Returns a subset of a varray objects

\item \code{dimnames<-}: Replacement method for dimnames (stops with error)

\item \code{dim<-}: Replacement method for dim (stops with error)

\item \code{length<-}: Replacement method for length (stops with error)

\item \code{mode<-}: Replacement method for mode (stops with error)

\item \code{[<-}: Replacement method for subset (stops with error)

\item \code{vstorage.mode}: Return the storage mode of the contents of the varray

\item \code{vstorage.mode<-}: Change the storage mode of the contents of the varray
}}

