% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/add.data.Matrix.R
\name{add.data}
\alias{add.data}
\alias{add.data.Matrix}
\alias{add.data.default}
\alias{add.data.matrix}
\title{Add data to an object}
\usage{
add.data(x, ...)

\method{add}{data.default}(x, ...)

\method{add}{data.Matrix}(x, newdata, need.dimnames = list(NULL, NULL),
  keep.ordered = TRUE, clear.dim = NULL, ..., envir = NULL)

\method{add}{data.matrix}(x, newdata, need.dimnames = list(NULL, NULL),
  keep.ordered = TRUE, clear.dim = NULL, ..., envir = NULL)
}
\arguments{
\item{x}{An object to add data to, either specified by name or by value.
If specified by name (as a character vector), the object is changed in place (i.e., the function will have side effects)}

\item{newdata}{New data to incorporate in the object.
Should have the same number of dimensions as the object
being updated (i.e.,
\code{length(dim(x))==length(dim(newdata))}).  Must have
dimnames.}

\item{need.dimnames}{Dimension names that should be included in the updated object.}

\item{keep.ordered}{Logical.  Specifies which dimensions should be
kept ordered.  Can be a single element or a vector with
length equal to the number of dimensions of
\code{object}.}

\item{clear.dim}{numeric: dimension indices for which to
clear all values when present, e.g., with clear.dim=1,
all rows that match a row supplied in newdata are cleared}

\item{envir}{where to find the object, or where to create it if it does not already exist}

\item{\dots}{Not used, but needed because \code{add.data()} could be a generic.}
}
\value{
The altered object \code{x}.  If \code{x} was the
name of an object, then the altered object is also
updated in place.
}
\description{
Update the contents of a Matrix or matrix object, adding new dimension
indices if necessary.
}
\details{
Can be used in multiple ways, either called as a generic or calling the method directly (to create the object if it does not already exist):
\itemize{
\item
add.data(x, newdata): adds newdata to existing object x and returns the modified x (no side effects)
\item
add.data.Matrix(x, newdata): adds newdata to existing Matrix object x (no side effects)
\item
add.data('x', newdata): adds newdata to existing object named 'x' and saves the modified x (has side effects)
\item
add.data.Matrix('x', newdata): adds newdata to existing Matrix object named 'x' and saves the modified object in 'x' OR if 'x' doesn't exist, creates new Matrix object with contents newdata and saves it in 'x' (has side effects)
}
}
\section{Methods (by class)}{
\itemize{
\item \code{data.default}: Default method that adds data to a object specified by a character name

\item \code{data.Matrix}: Add data to a Matrix object

\item \code{data.matrix}: Add data to an ordinary matrix object
}}
\note{
Not really closely related to \code{varray} objects, but supplied here
 as a useful analogue to \code{\link{add.tsdata.varray}}.
}
\examples{
x <- cbind(A=c(a=1))
add.data.matrix('x', cbind(B=c(b=2)))
x
}

