% Generated by roxygen2 (4.0.0): do not edit by hand
\name{add.tsdata}
\alias{add.tsdata}
\alias{add.tsdata.default}
\alias{add.tsdata.sparsetsmat}
\title{Add data to an object}
\usage{
add.tsdata(x, ...)

\method{add.tsdata}{default}(x, ...)

\method{add.tsdata}{sparsetsmat}(x, newdata, ...,
  sort.ids = non.null(x$sort.ids, FALSE), drop.unneeded.dates = TRUE,
  direct.df = TRUE)
}
\arguments{
\item{x}{An object to add time-series data to, either specified by name or by value.
If specified by name (as a character vector), the object is changed in place.}

\item{newdata}{The new data to add to the object.  Can be a dataframe, a sparsetsmat object, or matrix data.}

\item{...}{additional arguments.}

\item{sort.ids}{Same as for \code{sparsetsmat}.}

\item{drop.unneeded.dates}{Same as for \code{sparsetsmat}.}

\item{direct.df}{If TRUE, treat newdata as a standard
data.frame representation of a sparse tsmat object; i.e.,
dates are in column 1, ids in column 2, and values in
column 3.  If FALSE, \code{sparsetsmat(newdata, ...)} is
used to convert \code{newdata} to a sparsetsmat object.}
}
\description{
Add data to an object
}
\details{
Data is added to a sparsetsmat object by
converting both to data.frames, rbinding, and then
converting back to sparsetsmat objects.
}

