% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/add.tsdata.sparsetsmat.R
\name{add.tsdata}
\alias{add.tsdata}
\alias{add.tsdata.default}
\alias{add.tsdata.sparsetsmat}
\title{Add data to an object}
\usage{
add.tsdata(x, ...)

\method{add.tsdata}{default}(x, ..., pos = NULL)

\method{add.tsdata}{sparsetsmat}(x, newdata, ..., pos = 1,
  sort.ids = non.null(x$sort.ids, FALSE), drop.unneeded.dates = TRUE,
  shadow = FALSE, fullmask = TRUE, direct.df = TRUE)
}
\arguments{
\item{x}{An object to add time-series data to, either specified by name or by value.
If specified by name (as a character vector), the object is changed in place.}

\item{...}{additional arguments.}

\item{pos}{the search path position to find the variable}

\item{newdata}{The new data to add to the object.  Can be
a dataframe, a sparsetsmat object, or matrix data.  If
matrix data, explicit NA's 'stop' a prior value from
propagating forward, but missing columns do not.}

\item{sort.ids}{Same as for \code{sparsetsmat}.}

\item{drop.unneeded.dates}{Same as for \code{sparsetsmat}.}

\item{shadow}{If TRUE, drop all data in x that is preceeded by data in newdata (i.e., newdata casts a shadow over x)}

\item{fullmask}{If TRUE, drop all data in x that is overlapped by the rectangle of data in newdata (i.e., newdata casts a shadow over x)}

\item{direct.df}{If TRUE, treat newdata as a standard
data.frame representation of a sparse tsmat object; i.e.,
dates are in column 1, ids in column 2, and values in
column 3.  If FALSE, \code{sparsetsmat(newdata, ...)} is
used to convert \code{newdata} to a sparsetsmat object.}
}
\description{
Add data to an object
}
\details{
Data is added to a sparsetsmat object by
converting both to data.frames, rbinding, and then
converting back to sparsetsmat objects.
}

