% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/varray.R
\name{rm.varray}
\alias{rm.varray}
\title{Remove a varray and its component objects}
\usage{
rm.varray(x, list = NULL)
}
\arguments{
\item{x}{The name of a \code{varray} object to remove, either quoted or unquoted.}

\item{list}{A character vector of names of varray objects to be removed.}
}
\value{
Returns the vector of names of objects as an invisible object.
}
\description{
Remove a varray and its component objects
}
\details{
Removes all the component objects of a varray,
then removes the varray object itself.
}

