% Generated by roxygen2 (4.0.0): do not edit by hand
\name{as.matrix.sparsetsmat}
\alias{as.array.sparsetsmat}
\alias{as.matrix.sparsetsmat}
\alias{colnames.sparsetsmat}
\alias{dim.sparsetsmat}
\alias{dimnames.sparsetsmat}
\alias{ncol.sparsetsmat}
\alias{nrow.sparsetsmat}
\alias{print.sparsetsmat}
\alias{rownames.sparsetsmat}
\title{Standard matrix/array methods for sparsetsmat objects}
\usage{
\method{as.matrix}{sparsetsmat}(x, ...)

\method{as.array}{sparsetsmat}(x, ...)

\method{dim}{sparsetsmat}(x)

\method{dimnames}{sparsetsmat}(x)

\method{nrow}{sparsetsmat}(x)

\method{ncol}{sparsetsmat}(x)

\method{rownames}{sparsetsmat}(x)

\method{colnames}{sparsetsmat}(x)

\method{print}{sparsetsmat}(x, ...)
}
\arguments{
\item{x}{a sparsetsmat object}

\item{...}{additional arguments}
}
\description{
as.matrix: Convert a sparsetsmat object to an ordinary (non-sparse) matrix

as.array: Convert a sparsetsmat object to an ordinary (non-sparse) matrix

dim: return \code{dim} of a sparsetsmat

dimnames: return \code{dimnames} of a sparsetsmat

nrow: return number of rows (virtual) of a sparsetsmat

ncol: return number of cols of a sparsetsmat

rownames: return rownames of a sparsetsmat

colnames: return colnames of a sparsetsmat
}

