% Generated by roxygen2 (4.0.0): do not edit by hand
\name{varray.methods}
\alias{[.varray}
\alias{[<-.varray}
\alias{as.array.varray}
\alias{as.matrix.varray}
\alias{dim.varray}
\alias{dim<-.varray}
\alias{dimnames.varray}
\alias{dimnames<-.varray}
\alias{length<-.varray}
\alias{mode.varray}
\alias{mode<-.varray}
\alias{print.varray}
\alias{storage.mode.varray}
\alias{varray.methods}
\alias{vstorage.mode}
\alias{vstorage.mode.varray}
\alias{vstorage.mode<-}
\alias{vstorage.mode<-.varray}
\title{Common matrix and array methods for varray objects}
\usage{
varray.methods()

\method{as.array}{varray}(x, ..., fill = x$fill)

\method{as.matrix}{varray}(x, ..., fill = x$fill)

\method{print}{varray}(x, ...)

\method{dim}{varray}(x)

\method{dimnames}{varray}(x)

\method{mode}{varray}(x)

\method{storage.mode}{varray}(x)

\method{[}{varray}(x, ..., drop = TRUE)

\method{dimnames}{varray}(x) <- value

\method{dim}{varray}(x) <- value

\method{length}{varray}(x) <- value

\method{mode}{varray}(x) <- value

\method{[}{varray}(x, i, j, ...) <- value

vstorage.mode(x, deep = FALSE)

\method{vstorage.mode}{varray}(x, deep = FALSE)

vstorage.mode(x) <- value

\method{vstorage.mode}{varray}(x) <- value
}
\arguments{
\item{x}{a varray object}

\item{i,j,\dots}{indexing arguments, treated as for ordinary array indexing}

\item{drop}{should dimensions with extent 1 in the result be dropped?}

\item{fill}{fill value to use in subsetting}

\item{value}{new values to use}

\item{deep}{should individual components be inspected to determine the result?}
}
\description{
Common matrix and array methods for varray objects
}
\section{Methods (by generic)}{
\itemize{
\item \code{as.array}: Convert a varray to an ordinary array

\item \code{as.matrix}: Convert a varray to an ordinary matrix

\item \code{print}: Print info about a varray object (but not the whole object)

\item \code{dim}: Return dimension size of a varray object

\item \code{dimnames}: Return dimension names of a varray object

\item \code{mode}: Return 'mode' of data in a varray object

\item \code{storage.mode}: Return 'storage.mode' of data in a varray object

\item \code{[}: Returns a subset of a varray objects

\item \code{dimnames<-}: Replacement method for dimnames (stops with error)

\item \code{dim<-}: Replacement method for dim (stops with error)

\item \code{length<-}: Replacement method for length (stops with error)

\item \code{mode<-}: Replacement method for mode (stops with error)

\item \code{[<-}: Replacement method for subset (stops with error)

\item \code{vstorage.mode}: Return the storage mode of the contents of the varray

\item \code{vstorage.mode<-}: Change the storage mode of the contents of the varray
}}

