% Generated by roxygen2 (4.0.0): do not edit by hand
\name{write.pdfmark}
\alias{write.pdfmark}
\title{Write an R structure of pdf bookmarks to a pdfmark file}
\usage{
write.pdfmark(marks, file = stdout(), pageformat = "")
}
\arguments{
\item{marks}{a dataframe of pdf bookmarks (see \code{\link{pdfmark}})}

\item{file}{a filename or connection to write to}

\item{pageformat}{if non-empty, this character string format is used for appending page numbers to the titles in the table of contents, e.g., ' [\%s]'.}
}
\description{
Internal utility function to create an intermediate file of PDF bookmarks
}
\seealso{
pdfmark
}

